Move Plugin.enable and Plugin.disable into PluginManager for a better java API.
