+++ backend/mirai-console/src/main/kotlin/net/mamoe/mirai/console/internal/plugin/PluginManagerImpl.kt
@@ -61,7 +61,7 @@ internal object PluginManagerImpl : PluginManager {
 
     init {
         MiraiConsole.coroutineContext[Job]!!.invokeOnCompletion {
-            plugins.forEach(Plugin::disable)
+            plugins.forEach { it.disable() }
         }
     }
